requirements.txt
django
gunicorn
django-heroku
django-allauth
django-crispy-forms
Pillow
djangorestframework
whitenoise